avant dernier commit

+++ src/main/java/project/dao/AccountDao.java
@@ -2,6 +2,6 @@ package project.dao;
 
 import project.model.Account;
 
-public interface AccountDao extends GenericDao<Account, Integer> {
+public interface AccountDao extends GenericDao<Account, Long> {
 
 }
