+++ src/main/java/project/dao/impl/AccountDaoImpl.java
@@ -5,6 +5,6 @@ import project.dao.AccountDao;
 import project.model.Account;
 
 @Repository
-public class AccountDaoImpl extends GenericDaoImpl<Account, Integer> implements AccountDao {
+public class AccountDaoImpl extends GenericDaoImpl<Account, Long> implements AccountDao {
 
 }
